+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 - Charge order matching fee through the margin
+- Improve wallet balance design
 
 ## [1.4.2] - 2023-10-18